Update docs and cleanup
