+++ rpgJavaInterpreter-core/src/main/kotlin/com/smeup/rpgparser/execution/MainExecutionContext.kt
@@ -19,15 +19,10 @@ package com.smeup.rpgparser.execution
 import com.smeup.dbnative.manager.DBFileFactory
 import com.smeup.rpgparser.experimental.ExperimentalFeaturesFactory
 import com.smeup.rpgparser.interpreter.*
-import com.smeup.rpgparser.logging.LogChannel
+import com.smeup.rpgparser.logging.LoggingContext
 import com.smeup.rpgparser.parsing.facade.CopyBlocks
-import java.text.DecimalFormat
 import java.util.*
 import java.util.concurrent.atomic.AtomicInteger
-import kotlin.collections.HashMap
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.nanoseconds
-import kotlin.time.DurationUnit
 
 /**
  * Execution context allows to propagate, in simple and safe mode, some useful information, that could be
@@ -39,7 +34,6 @@ object MainExecutionContext {
 
     // default values in case jariko is not called from the command line program
     private val context: ThreadLocal<Context> by lazy { ThreadLocal<Context>() }
-    private val loggingContext: ThreadLocal<LoggingContext> by lazy { ThreadLocal<LoggingContext>() }
     private val noContextIdProvider: AtomicInteger by lazy { AtomicInteger() }
     private val noContextAttributes: MutableMap<String, Any> by lazy { mutableMapOf<String, Any>() }
     private val noConfiguration: Configuration by lazy { Configuration() }
@@ -87,9 +81,6 @@ object MainExecutionContext {
                         systemInterface = systemInterface
                     )
                 )
-                loggingContext.set(
-                    LoggingContext()
-                )
             }
             return mainProgram.runCatching {
                 invoke(context.get())
@@ -102,7 +93,6 @@ object MainExecutionContext {
             if (isRootContext) {
                 context.get()?.dbFileFactory?.close()
                 context.remove()
-                loggingContext.remove()
             }
         }
     }
@@ -115,7 +105,7 @@ object MainExecutionContext {
     /**
      * @return logging context
      */
-    fun getLoggingContext() = loggingContext.get() ?: null
+    fun getLoggingContext() = context.get()?.logging
 
     /**
      * @return a new unique identifier
@@ -210,6 +200,7 @@ data class Context(
     val attributes: MutableMap<String, Any> = mutableMapOf<String, Any>(),
     val idProvider: AtomicInteger = AtomicInteger(),
     val configuration: Configuration,
+    val logging: LoggingContext = LoggingContext(),
     val memorySliceMgr: MemorySliceMgr? = null,
     val programStack: Stack<RpgProgram> = Stack<RpgProgram>(),
     val systemInterface: SystemInterface,
@@ -235,116 +226,4 @@ data class ParsingProgram(val name: String) {
     var copyBlocks: CopyBlocks? = null
     var sourceLines: List<String>? = null
     val attributes: MutableMap<String, Any> = mutableMapOf()
-}
-
-class LoggingContext {
-    private val timeUsageByStatement: HashMap<String, UsageMeasurement> = hashMapOf()
-    private val symbolTableTimeUsage: EnumMap<SymbolTableAction, UsageMeasurement> = EnumMap(SymbolTableAction::class.java)
-    private var renderingTimeMeasurement = UsageMeasurement.new()
-    private var expressionTimeMeasurement = UsageMeasurement.new()
-
-    private val initTimestamp = System.nanoTime()
-    private val percentageFormatter = DecimalFormat("##%")
-
-    private val totalTime
-        get() = (System.nanoTime() - initTimestamp).nanoseconds
-
-    enum class SymbolTableAction {
-        INIT,
-        LOAD,
-        STORE
-    }
-
-    data class UsageMeasurement(val duration: Duration, val hit: Long) {
-        companion object {
-            fun new(): UsageMeasurement = UsageMeasurement(
-                duration = Duration.ZERO,
-                hit = 0
-            )
-        }
-    }
-
-    fun recordRenderingDuration(executionTime: Duration) {
-        renderingTimeMeasurement = UsageMeasurement(
-            duration = renderingTimeMeasurement.duration + executionTime,
-            hit = renderingTimeMeasurement.hit + 1
-        )
-    }
-
-    fun recordExpressionDuration(executionTime: Duration) {
-        expressionTimeMeasurement = UsageMeasurement(
-            duration = renderingTimeMeasurement.duration + executionTime,
-            hit = renderingTimeMeasurement.hit + 1
-        )
-    }
-
-    fun recordSymbolTableDuration(action: SymbolTableAction, executionTime: Duration) {
-        val entry = symbolTableTimeUsage.getOrPut(action) { UsageMeasurement.new() }
-        symbolTableTimeUsage[action] = UsageMeasurement(
-            duration = entry.duration + executionTime,
-            hit = entry.hit + 1
-        )
-    }
-
-    fun recordStatementDuration(name: String, executionTime: Duration) {
-        val entry = timeUsageByStatement.getOrPut(name) { UsageMeasurement.new() }
-        timeUsageByStatement[name] = UsageMeasurement(
-            duration = entry.duration + executionTime,
-            hit = entry.hit + 1
-        )
-    }
-
-    fun generateTimeUsageByStatementReportEntries(source: LogSourceData): List<LazyLogEntry> {
-        return timeUsageByStatement.toList().map {
-            val statementName = it.first
-            val duration = it.second.duration
-            val timePercentage = percentageFormatter.format(duration / totalTime)
-            val hit = it.second.hit
-            val average = (duration.inWholeNanoseconds / hit).nanoseconds
-
-            val entry = LogEntry(source, LogChannel.ANALYTICS.getPropertyName(), "STMT TIME")
-            LazyLogEntry(entry) { sep ->
-                "$statementName$sep${duration.toString(DurationUnit.MICROSECONDS)}$sep$timePercentage${sep}$hit${sep}avg. ${average.toString(DurationUnit.MICROSECONDS)}"
-            }
-        }
-    }
-
-    fun generateSymbolTableTimeUsageReportEntries(source: LogSourceData): List<LazyLogEntry> {
-        return symbolTableTimeUsage.toList().map {
-            val action = it.first
-            val duration = it.second.duration
-            val timePercentage = percentageFormatter.format(duration / totalTime)
-            val hit = it.second.hit
-            val average = (duration.inWholeNanoseconds / hit).nanoseconds
-
-            val entry = LogEntry(source, LogChannel.ANALYTICS.getPropertyName(), "SYMTBL TIME")
-            LazyLogEntry(entry) { sep ->
-                "${action.name}$sep${duration.toString(DurationUnit.MICROSECONDS)}$sep$timePercentage${sep}$hit${sep}avg. ${average.toString(DurationUnit.MICROSECONDS)}"
-            }
-        }
-    }
-
-    fun generateExpressionReportEntry(source: LogSourceData): LazyLogEntry {
-        val duration = expressionTimeMeasurement.duration
-        val timePercentage = percentageFormatter.format(duration / totalTime)
-        val hit = expressionTimeMeasurement.hit
-        val average = (duration.inWholeNanoseconds / hit).nanoseconds
-
-        val entry = LogEntry(source, LogChannel.ANALYTICS.getPropertyName(), "EXPR TIME")
-        return LazyLogEntry(entry) { sep ->
-            "${duration.toString(DurationUnit.MICROSECONDS)}$sep$timePercentage$sep$hit${sep}avg. ${average.toString(DurationUnit.MICROSECONDS)}"
-        }
-    }
-
-    fun generateLogTimeReportEntry(source: LogSourceData): LazyLogEntry {
-        val duration = renderingTimeMeasurement.duration
-        val timePercentage = percentageFormatter.format(duration / totalTime)
-        val hit = renderingTimeMeasurement.hit
-        val average = (duration.inWholeNanoseconds / hit).nanoseconds
-
-        val entry = LogEntry(source, LogChannel.ANALYTICS.getPropertyName(), "LOG TIME")
-        return LazyLogEntry(entry) { sep ->
-            "${duration.toString(DurationUnit.MICROSECONDS)}$sep$timePercentage$sep$hit${sep}avg. ${average.toString(DurationUnit.MICROSECONDS)}"
-        }
-    }
 }
